first draft: extract parseRange()
